third_party/blink/web_tests/webaudio/internals/scriptprocessornode-premature-death-expected.txt

CONSOLE WARNING: The ScriptProcessorNode is deprecated. Use AudioWorkletNode instead. (https://bit.ly/audio-worklet)
CONSOLE WARNING: The ScriptProcessorNode is deprecated. Use AudioWorkletNode instead. (https://bit.ly/audio-worklet)
Tests that a script processor node is not prematurely GCed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing with explicitly keeping a reference to the script processor node alive.
PASS wasCollectedPrematurely is false
PASS wasCalled is true
Testing without explicitly keeping a reference to the script processor node alive.
PASS wasCollectedPrematurely is false
PASS wasCalled is true
PASS successfullyParsed is true

TEST COMPLETE
